api_requirements.txt
Flask==2.0.2
Flask-Cors==3.0.10
gunicorn==20.1.0
py2neo==2021.2.4
python-dotenv==0.19.1
